requirements.txt
fastmcp>=0.1.0
duckduckgo-search>=4.1.1